chore: Add dependencies and configuration
Decisions for discussion/approval:

1. I've used the conventional commit specification for this commit
message.
I'm keen to ensure a clean git history and like the conventional commit
format. However, I'm open to discussion on what commit message format we
should use.
link: https://www.conventionalcommits.org/en/v1.0.0/

2. Using pip-tools for dependency management
I like pip-tools as it documents the parent package of each dependency
installed in the project. In my opinion this makes dependency management
easier in the longer term when adding or removing packages. pip tools
also contains a `pip-sync` command that updates the virtual environment
to match the `requirements.txt` or `requirements-dev.txt` files,
simplifying installing or uninstalling packages.
Link: https://github.com/jazzband/pip-tools

3. Using separate requirement files for production and development
dependencies
I've created separate requirement.in and requirements-dev.in files and
used `pip-compile` to create the repective `requirements.txt` and
`requirements-dev.txt` files. I realise this adds 4 files for dependency
management which might not be to everyone's taste. As an alternative pip
tools provides the option of managing dependencies in a `pyproject.toml` file.

4. I've assumed we will use Python 3.11
If changing Python version `.pre-commit-config.yaml` and
`pyproject.toml` will require updating:
	`.pre-commit-config.yaml` references python 3.11 as the default
	python version for hooks
	`pyproject.toml` references python 3.11 as the target version
	for black

Resolves #3