feat: add home app, add minimal index and calendar views to build upon
My current thought is that _at minimum_ we need two pages:
a) One landing page where visitors to the Reading Tracker arrive
- This will include some basic info what this page is about and also Sign Up & Login Buttons
b) A "Calendar" page that is displayed once a reader is logged in
- This will include the "streak calendar", as well as a button to say "I was active today"
- (Also a "Sign out" button, and – possibly a bit later – a way to make adjustments, if a day was accidentally forgotten to be added.)

This commit adds the base skeleton. Next, I intend to add user management via django-allauth, as well as add a Navbar and some simple styling via Tailwind/DaisyUI (using CDN for now to start).