feat: set sensible low-friction settings for django-allauth authentication flow
- Keep the friction as low as possible. Single password field. No authentication e-mail needed to start.
- Allow for sign-up with only username and pw.
- Use custom sign-up form do remove the e-mail field on the sign-up page

Thought: We could also consider restricting sign-up for new users and manually adding new people that join the reading club. But only if the "wide open" sign-up turns out to be a problem. To begin we can just try to keep it open for direct sign-up and see what happens.. And maybe later we want to add e-mail back in to have motivational messages, streak-celebrations, or similar :D